DOU with captured_list: Large change on captura_dou, http-request and parametrize-API-requests
To make DOU articles' capture more robust, we changed both the DOUTOR system of posting articles
to Slack and the main capturing system "parametrize-API-requests" + "http-requests" (along with
necessary dynamoDB tables) to register in 'dou_capture_urls' (posting to slack) and
'douDB_capture_urls' (making DOU DataBase) the articles from the current day that were already
downloaded. The system was tested with the live DOU capture, other DOU captures (that do not use
this system or config data in dynamoDB config table), and "fill missing 56 senadores proposicoes"
or something like that.