add-to-bigquery-slave with recursive data type replacement / requests without certificate as fallback
When setting the data types to string in add-to-bigquery-slave, do it recursively in case of structured data
(before this, weird data types such dates in structs were left as is. Now, it changes to string).

Due to a temporary problem in downloading DOU data, now http-request and capture_dou first try the regular
http request. If it fails with SSL error, try without asking for a certificate. This solved the problem
with downloading DOU data.